Merging the working version of KnowNet Application

+++ knownet-web/src/main/java/com/ncedu/knownetweb/LessonController.java
@@ -2,18 +2,24 @@ package com.ncedu.knownetweb;
 
 import com.ncedu.knownetimpl.model.LessonBody;
 import com.ncedu.knownetimpl.model.entity.Lesson;
+import com.ncedu.knownetimpl.model.entity.Role;
+import com.ncedu.knownetimpl.model.entity.User;
 import com.ncedu.knownetimpl.service.LessonService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
 import java.util.Optional;
 
 
 @Slf4j
 @RestController
+@CrossOrigin(origins = "http://localhost:4200")
 @RequestMapping("/lessons")
 public class LessonController {
     private final LessonService lessonService;
@@ -53,20 +59,28 @@ public class LessonController {
         log.debug("requested: learnRequests get    (teacherId = {})", teacherId);
         return ResponseEntity.ok().body(lessonService.findByTeacherId(teacherId));
     }
-    
+
     @DeleteMapping(value = "byId/{id}")
-    public ResponseEntity<String> deleteById(@PathVariable("id") Long id) {
+    public ResponseEntity<String> deleteById(@PathVariable("id") Long id, Authentication authentication) {
         log.debug("requested: lesson  delete (id = {})", id);
-        boolean deleted = lessonService.deleteById(id);
-        if (deleted) {
-            return ResponseEntity.ok().body("lesson with id = " + id + " was deleted");
-        } else {
+
+        Optional<Lesson> lesson = lessonService.findById(id);
+        if (!lesson.isPresent()) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body("lesson with id = " + id + " does not exist");
         }
+
+        User user = (User)authentication.getPrincipal();
+        if (user.getRole() == Role.ADMIN || lesson.get().getTeacher().getLogin().equals(user.getLogin())) {
+            lessonService.deleteById(id);
+            return ResponseEntity.ok().body("lesson with id = " + id + " was deleted");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("");
+        }
     }
 
-    @PostMapping(value = "lesson")
+    @PreAuthorize("#lessonBody.teacherId == authentication.principal.id or hasRole('ADMIN')")
+    @PostMapping(value = "")
     public ResponseEntity<String> create(@RequestBody LessonBody lessonBody) {
         Lesson lesson = lessonService.makeFromBody(lessonBody);
         log.debug("requested: lesson  create (teacher_id = {}, tagId = {}, lessonName = {})",
@@ -81,9 +95,15 @@ public class LessonController {
         }
     }
 
-    @PutMapping(value = "lesson")
-    public ResponseEntity<String> update(@RequestBody LessonBody lessonBody) {
+    @PutMapping(value = "")
+    public ResponseEntity<String> update(@RequestBody LessonBody lessonBody, Authentication authentication) {
         Lesson lesson = lessonService.makeFromBody(lessonBody);
+
+        User user = (User)authentication.getPrincipal();
+        if (!(user.getRole() == Role.ADMIN) && !lesson.getTeacher().getLogin().equals(user.getLogin())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("");
+        }
+
         Long id = lesson.getId();
         log.debug("requested: lesson  update (id = {})", id);
         boolean updated = lessonService.update(lesson);
